# Conflicts:
#	src/main/java/com/warehouse/myshop/product/repository/ProductRepository.java
#	src/main/resources/application-local.yml

+++ src/test/java/com/warehouse/myshop/product/ProductTestBase.java
@@ -63,7 +63,7 @@ public abstract class ProductTestBase {
                 .name("test")
                 .articleNumber("te-st-article")
                 .description("test description")
-                .price(50.5)
+                .price(BigDecimal.valueOf(50.5))
                 .quantity(5)
                 .build();
     }
@@ -74,7 +74,7 @@ public abstract class ProductTestBase {
                 .name("test")
                 .articleNumber("te-st-article")
                 .description("test description")
-                .price(50.5)
+                .price(BigDecimal.valueOf(50.5))
                 .quantity(10)
                 .build();
     }
